Bump min SDK version to 10.9 and suppress deprecation warnings
This implies using a modern C++ library (libc++) which can use C++11
features we rely on for cross compiling to Android. This suppresses
deprecation warnings in base and skia as those libraries need to work
in environments that target a min sdk version of 10.6.

R=cstout@chromium.org, tonyg@chromium.org

Review URL: https://codereview.chromium.org/1318153005 .